Monolithic commit; working on multiswarm
