+++ package.json
@@ -25,6 +25,7 @@
     "google-protobuf": "^3.6.1",
     "hash.js": "^1.1.5",
     "isomorphic-ws": "^4.0.1",
+    "lodash": "^4.17.11",
     "web3": "^1.0.0-beta.51",
     "ws": "^6.0.0"
   }
