+++ src/swarmClient/5_broadcast_layer.js
@@ -14,15 +14,15 @@
 
 
 const assert = require('assert');
-const database_pb = require('../proto/database_pb');
-const bluzelle_pb = require('../proto/bluzelle_pb');
-const status_pb = require('../proto/status_pb');
+const bluzelle_pb = require('../../proto/bluzelle_pb');
+const database_pb = require('../../proto/database_pb');
+const status_pb = require('../../proto/status_pb');
 const {BroadcastSocket} = require('./1_connection_layer');
 
 
 module.exports = class Broadcast {
 
-    constructor({p2p_latency_bound, connection_layer, onIncomingMsg, log, onOutgoingMsg}) {
+    constructor({p2p_latency_bound, peerslist, connection_layer, onIncomingMsg, log, onOutgoingMsg}) {
 
         this.onIncomingMsg = onIncomingMsg;
         this.onOutgoingMsg = onOutgoingMsg;
@@ -32,7 +32,9 @@ module.exports = class Broadcast {
         this.p2p_latency_bound = p2p_latency_bound;
         this.connection_layer = connection_layer;
 
-        this.peers = [];
+        this.peerslist = peerslist;
+
+
         this.sockets;
 
         this.timeout = p2p_latency_bound * 15;
@@ -42,24 +44,21 @@ module.exports = class Broadcast {
     }
 
 
-    sendOutgoingMsg(bzn_envelope) {
+    sendOutgoingMsg(bzn_envelope, msg) {
 
         assert(bzn_envelope instanceof bluzelle_pb.bzn_envelope);
 
-        if(bzn_envelope.hasDatabaseMsg()) {
-
-            const database_msg = database_pb.database_msg.deserializeBinary(new Uint8Array(bzn_envelope.getDatabaseMsg()));
-
+        if(msg instanceof database_pb.database_msg) {
 
             // Do not broadcast quickreads
 
-            if(database_msg.hasQuickRead()) {
-                this.onOutgoingMsg(bzn_envelope);
+            if(msg.hasQuickRead()) {
+                this.onOutgoingMsg(bzn_envelope, msg);
                 return;
             }
 
 
-            const nonce = database_msg.getHeader().getNonce();
+            const nonce = msg.getHeader().getNonce();
             setTimeout(() => {
                     
                 const fn = this.timeoutFns.get(nonce);
@@ -72,11 +71,11 @@ module.exports = class Broadcast {
             // This gets overwritten when the messages comes back
             // so a properly-responded message does not execute a broadcast
 
-            this.timeoutFns.set(nonce, () => this.broadcast(bzn_envelope)); 
+            this.timeoutFns.set(nonce, () => this.broadcast(bzn_envelope, msg)); 
 
         } 
 
-        this.onOutgoingMsg(bzn_envelope);
+        this.onOutgoingMsg(bzn_envelope, msg);
 
     }
 
@@ -84,13 +83,6 @@ module.exports = class Broadcast {
     sendIncomingMsg(bzn_envelope) {
 
         assert(bzn_envelope instanceof bluzelle_pb.bzn_envelope);
-
-        if(bzn_envelope.hasStatusResponse()) {
-
-            const status_response = status_pb.status_response.deserializeBinary(new Uint8Array(bzn_envelope.getStatusResponse()));
-            this.peers = JSON.parse(status_response.toObject().moduleStatusJson).module[0].status.peer_index;
-
-        }
 
 
          if(bzn_envelope.hasDatabaseResponse()) {
@@ -113,25 +105,23 @@ module.exports = class Broadcast {
     }
 
 
-    broadcast(bzn_envelope) {
+    broadcast(bzn_envelope, msg) {
 
         this.log && this.log('Rebroadcasting msg');
-
-        assert(this.peers.length > 0);
 
 
         if(!this.sockets) { 
 
-            this.sockets = this.peers.map(peer => new BroadcastSocket({
+            this.sockets = Object.values(this.peerslist).map(peer => new BroadcastSocket({
                 onmessage: m => this.connection_layer.sendIncomingMsg(m),
-                entry: 'ws://' + peer.host + ':' + peer.port,
+                entry: 'ws://' + peer.nodeHost + ':' + peer.nodePort,
                 connection_pool: this.connection_layer.connection_pool,
                 log: this.log
             }));
 
         }
 
-        this.sendOutgoingMsg(bzn_envelope);
+        this.sendOutgoingMsg(bzn_envelope, msg);
 
     }
