+++ src/swarmClient/6_redirect_layer.js
@@ -13,8 +13,9 @@
 // along with this program. If not, see <http://www.gnu.org/licenses/>.
 
 
-const database_pb = require('../proto/database_pb');
-const bluzelle_pb = require('../proto/bluzelle_pb');
+const bluzelle_pb = require('../../proto/bluzelle_pb');
+const database_pb = require('../../proto/database_pb');
+const status_pb = require('../../proto/status_pb');
 
 const assert = require('assert');
 
@@ -28,12 +29,12 @@ module.exports = class Redirect {
 
     }
 
-    sendOutgoingMsg(bzn_envelope) {
+    sendOutgoingMsg(bzn_envelope, msg) {
 
         assert(bzn_envelope instanceof bluzelle_pb.bzn_envelope);
 
         // Pass through
-        this.onOutgoingMsg(bzn_envelope);
+        this.onOutgoingMsg(bzn_envelope, msg);
 
     }
