+++ src/swarmClient/test/main.test.js
@@ -1,3 +1,17 @@
+// Copyright (C) 2019 Bluzelle
+//
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU Affero General Public License, version 3,
+// as published by the Free Software Foundation.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU Affero General Public License for more details.
+//
+// You should have received a copy of the GNU Affero General Public License
+// along with this program. If not, see <http://www.gnu.org/licenses/>.
+
 const {bluzelle, version} = require('../main');
 const assert = require('assert');
 const {pub_from_priv, random_key} = require('../ecdsa_secp256k1');
