--- test-daemon/configs/bluzelle0.json
@@ -1,13 +1,0 @@
-{
-  "listener_address" : "127.0.0.1",
-  "listener_port" : 50000,
-  "http_port" : 8080,
-  "ethereum" : "0xddbd2b932c763ba5b1b7ae3b362eac3e8d40121a",
-  "ethereum_io_api_token" : "*****************************",
-  "bootstrap_file" : "./peers.json",
-  "uuid" : "60ba0788-9992-4cdb-b1f7-9f68eef52ab9",
-  "debug_logging" : true,
-  "log_to_stdout" : true,
-  "use_pbft": true,
-  "crypto_enabled_outgoing": false
-}
