--- test-daemon/configs/bluzelle1.json
@@ -1,13 +1,0 @@
-{
-  "listener_address" : "127.0.0.1",
-  "listener_port" : 50001,
-  "http_port" : 8081,
-  "ethereum" : "0xddbd2b932c763ba5b1b7ae3b362eac3e8d40121a",
-  "ethereum_io_api_token" : "*****************************",
-  "bootstrap_file" : "./peers.json",
-  "uuid" : "c7044c76-135b-452d-858a-f789d82c7eb7",
-  "debug_logging" : true,
-  "log_to_stdout" : true,
-  "use_pbft": true,
-  "crypto_enabled_outgoing": false
-}
