--- test-daemon/scripts/kill-daemon.sh
@@ -1,3 +1,0 @@
-#!/usr/bin/env bash
-
-pkill -2 swarm
