--- test-daemon/scripts/run-daemon.sh
@@ -1,8 +1,0 @@
-#!/usr/bin/env bash
-
-if [ "$1" != "" ]; then
-    cd ../daemon-build/output; ./swarm -c $1
-else
-    cd ../daemon-build/output; ./swarm
-fi
-
